新增 進度條功能

+++ src/main/java/com/startrip/reviews/service/ReviewService.java
@@ -7,7 +7,7 @@ import com.startrip.reviews.model.ReviewBean;
 public interface ReviewService {
 	List<ReviewBean> getAllReviews();
 	List<ReviewBean> getReviewsByHotelId(int hotelId);
-	List<Integer> getRankByHotelId(int hotelId);
+	List<Long> getRankByHotelId(int hotelId);
 	
 	void addReview(ReviewBean review);
